# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AuthResponseDTO:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  domain.RegisterDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  domain.UserResponseDTO:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: suporte@runter.com
    name: Suporte da API
  description: API RESTful em Go com autenticação JWT, PostgreSQL e Chi Router.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Runter Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Valida o email e senha do usuário e retorna o token JWT de autorização.
      parameters:
      - description: Credenciais de acesso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponseDTO'
        "400":
          description: Requisição inválida
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Credenciais inválidas
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autenticar usuário
      tags:
      - Autenticação
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cadastra um novo usuário com nome, email e senha e retorna um token
        JWT de acesso.
      parameters:
      - description: Dados de registro do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AuthResponseDTO'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar novo usuário
      tags:
      - Autenticação
  /health:
    get:
      description: Retorna o status operacional da aplicação, timestamp UTC e versão
        da API.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Healthcheck da API
      tags:
      - Saúde
  /users/me:
    get:
      description: Retorna as informações do usuário atual extraídas a partir do token
        JWT enviado no cabeçalho Authorization.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserResponseDTO'
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obter perfil do usuário autenticado
      tags:
      - Usuários
securityDefinitions:
  BearerAuth:
    description: Digite 'Bearer ' seguido pelo seu token JWT obtido no login/registro.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"